feat(ui): modularize UI into 9 focused files with per-device configuration
Complete refactoring of the 1900+ line monolithic ui.rs into a modular
architecture following GTK3 best practices:

New UI Structure:
- ui/app.rs         - Main window, menu bar, tab management
- ui/audio_tab.rs   - Output/Input tab logic with per-device WirePlumber configs
- ui/advanced_tab.rs - Advanced mode with exclusive/professional settings
- ui/components.rs  - Reusable widgets (constrained combos, section boxes)
- ui/constants.rs   - Shared option arrays (sample rates, bit depths, etc.)
- ui/tab_types.rs   - TabType enum with input/output differentiation
- ui/dialogs.rs     - Error, success, and about dialogs
- ui/helpers.rs     - String cleaning utilities
- ui/preferences.rs - User preferences load/save

Per-Device Configuration:
- Apply WirePlumber configs using actual system device names
- Detect real device name from PipeWire/ALSA for default devices
- Generate device-specific configurations instead of wildcard patterns
- Call apply_wireplumber_device_config() alongside main settings apply

Configuration Inspector Improvements:
- Scan both PipeWire and WirePlumber directories
- Added ~/.config/wireplumber/main.lua.d/ scanning
- Added /usr/share/pipewire and /usr/share/wireplumber defaults
- Fixed double-dialog showing combined results once
- Added scan-in-progress guard to prevent duplicate scans
- Empty results are informational, not errors

CI/CD:
- Added Xvfb virtual display for GTK testing on GitHub Actions
- Updated rust-ci.yml with DISPLAY=:99.0 for headless GTK tests

Fixes:
- Restored app icon in About dialog with full icon path list
- Fixed impl ConfigInspectorTab block closing brace issue
- Removed duplicate create_wireplumber_config_new function
- Updated all callers for new 3-parameter function signatures
- Fixed unused HashMap import in audio.rs

This completes the v2.0 modularization milestone.
