+++ .github/workflows/rust-ci.yml
@@ -5,7 +5,7 @@ on:
     branches: [ stage, main ]
   pull_request:
     branches: [ stage, main ]
-  workflow_dispatch:   # allow manual runs (useful to trigger full integration)
+  workflow_dispatch:
     inputs:
       run_integration:
         description: 'Run PipeWire integration tests (requires suitable runner)'
@@ -23,9 +23,15 @@ jobs:
       - name: Install system dependencies (GTK, ALSA, PipeWire, WirePlumber)
         run: |
           sudo apt-get update -y
-          sudo apt-get install -y build-essential pkg-config libglib2.0-dev libgtk-3-dev libasound2-dev libpipewire-0.3-dev pipewire wireplumber
+          sudo apt-get install -y \
+            build-essential pkg-config \
+            libglib2.0-dev libgtk-3-dev \
+            libasound2-dev \
+            libpipewire-0.3-dev \
+            pipewire wireplumber \
+            xvfb  # Virtual framebuffer for GTK tests
 
-      - name: Set up Rust toolchain (actions-rs/toolchain)
+      - name: Set up Rust toolchain
         uses: actions-rs/toolchain@v1
         with:
           toolchain: stable
@@ -49,9 +55,28 @@ jobs:
       - name: cargo clippy (warn only)
         run: cargo clippy --all-targets --all-features
 
-      - name: Run tests
+      - name: Build (check compilation)
+        run: cargo build --workspace --all-features
+
+      - name: Run unit tests (no GTK/display needed)
         env:
           CI: "true"
           GITHUB_ACTIONS: "true"
+        run: cargo test --workspace --all-features --verbose -- --test-threads=1
+
+      - name: Run GTK tests with virtual display
+        env:
+          CI: "true"
+          GITHUB_ACTIONS: "true"
+          DISPLAY: ":99.0"
           RUN_PIPEWIRE_INTEGRATION: ${{ github.event.inputs.run_integration || 'false' }}
-        run: cargo test --workspace --all-features --verbose
+        run: |
+          # Start Xvfb virtual display
+          Xvfb :99 -screen 0 1280x1024x24 &
+          sleep 2
+          
+          # Run tests with virtual display
+          cargo test --workspace --all-features --verbose -- --test-threads=1
+          
+          # Clean up Xvfb
+          killall Xvfb || true
