+++ README.md
@@ -1,7 +1,8 @@
 # Pro Audio Config <div style="text-align: right">[![Rust CI](https://github.com/Peter-L-SVK/pro_audio_config/actions/workflows/rust-ci.yml/badge.svg)](https://github.com/Peter-L-SVK/pro_audio_config/actions/workflows/rust-ci.yml)</div>
 
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT) 
-[![Github All Releases](https://img.shields.io/github/downloads/Peter-L-SVK/pro_audio_config/total.svg)](https://github.com/Peter-L-SVK/pro_audio_config) <a href="https://buymeacoffee.com/leukanic.peter"><img src="https://www.buymeacoffee.com/assets/img/custom_images/orange_img.png" height="20px"></a>
+[![Top Language](https://img.shields.io/github/languages/top/Peter-L-SVK/pro_audio_config)](https://github.com/Peter-L-SVK/pro_audio_config)
+[![GitHub release (latest by date)](https://img.shields.io/github/v/release/Peter-L-SVK/pro_audio_config)](https://github.com/Peter-L-SVK/pro_audio_config/releases/latest)<a href="https://buymeacoffee.com/leukanic.peter"><img src="https://www.buymeacoffee.com/assets/img/custom_images/orange_img.png" height="20px"></a>
 
 A professional audio configuration tool for Linux systems that provides a simple graphical interface to manage PipeWire and ALSA audio settings. Finally, an easy way to configure sample rates, bit depths, and buffer sizes without digging through config files.
